routers were created. Models were separated from main. Refactored main.py
